Fixed bug in GRIP generated code for FilterContours--missing cast to elevate width and height from integer to float.

+++ 2017/Robot2017/src/frc2135/GripContoursPipeline.cpp
@@ -140,7 +140,7 @@ std::vector<std::vector<cv::Point> >* GripContoursPipeline::getfilterContoursOut
 			double solid = 100 * area / cv::contourArea(hull);
 			if (solid < solidity[0] || solid > solidity[1]) continue;
 			if (contour.size() < minVertexCount || contour.size() > maxVertexCount)	continue;
-			double ratio = bb.width / bb.height;
+			double ratio = (double)bb.width / (double)bb.height;
 			if (ratio < minRatio || ratio > maxRatio) continue;
 			output.push_back(contour);
 		}
